Rename: 컴포넌트 이름 변경
페이지 내 컴포넌트의 경우, 명확한 파일 식별을 위해  파일 이름 뒤에 Component를 붙이기로 함.
